app: extract reusable widgets/helper, enable lint rules, add test
- Move the three self-contained pieces out of the 1193-line home_page.dart into
  lib/widgets/: the pure wrapKeyTitle() text-wrapper (key_title.dart, now
  unit-testable), and the standalone KeepAlive and PlayPulse widgets. home_page
  imports them; behaviour is unchanged (verbatim move + rename of references).
- Add test/key_title_test.dart covering the wrapping invariants (<=12 chars/line,
  <=3 lines, ellipsis on overflow, break-before-capital).
- Enable a conservative set of project lint rules in analysis_options.yaml that
  match the existing style (single quotes, const, final locals, no print).
- Add reason comments to the two remaining bare catch(_) blocks in the controller.

Note: Flutter/Dart is not available in this environment; run `flutter analyze`
and `flutter test` locally to confirm. The deeper split of the stateful builder
methods (_grid/_key/_controls) is left as a follow-up to do with a compiler, as
it requires threading controller/callbacks and can't be verified here.
